# Optimize types.ParseTime and add benchmark test for it