--- a/types/scalar_types_test.go
+++ b/types/scalar_types_test.go
@@ -23,6 +23,65 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+var datesWithTz = []struct {
+	in  string
+	out time.Time
+}{
+	{in: "2018-10-28T04:00:10Z",
+		out: time.Date(2018, 10, 28, 4, 00, 10, 0, time.UTC)},
+	{in: "2018-10-28T04:00:10-00:00",
+		out: time.Date(2018, 10, 28, 4, 00, 10, 0, time.UTC)},
+	{in: "2018-05-30T09:30:10.5Z",
+		out: time.Date(2018, 5, 30, 9, 30, 10, 500000000, time.UTC)},
+	{in: "2018-05-30T09:30:10.5-00:00",
+		out: time.Date(2018, 5, 30, 9, 30, 10, 500000000, time.UTC)},
+	{in: "2018-05-30T09:30:10-06:00",
+		out: time.Date(2018, 5, 30, 9, 30, 10, 0, time.FixedZone("", -6*60*60))},
+	{in: "2018-05-28T14:41:57+30:00",
+		out: time.Date(2018, 5, 28, 14, 41, 57, 0, time.FixedZone("", 30*60*60))},
+}
+
+var datesWithoutTz = []struct {
+	in  string
+	out time.Time
+}{
+	{in: "2018-10-28T04:00:10",
+		out: time.Date(2018, 10, 28, 4, 00, 10, 0, time.UTC)},
+	{in: "2018-05-30T09:30:10.5",
+		out: time.Date(2018, 5, 30, 9, 30, 10, 500000000, time.UTC)},
+	{in: "2018",
+		out: time.Date(2018, 1, 1, 0, 0, 0, 0, time.UTC)},
+	{in: "2018-01",
+		out: time.Date(2018, 1, 1, 0, 0, 0, 0, time.UTC)},
+	{in: "2018-01-01",
+		out: time.Date(2018, 1, 1, 0, 0, 0, 0, time.UTC)},
+}
+
+var invalidDates = []string{
+	"abcd",
+	"12345",
+	"123456",
+	"1234567",
+	"12345678",
+	"123456789",
+	"1234567891",
+	"11111111111111111Z",
+	"111111111111111:11",
+	"1111-11-11T11:11111111:1",
+	"1111-11-11T11:11:1111:11",
+	"18-10-28T04:00:10Z",
+	"318-10-28T04:00:10",
+	"2018-110-28T04:00:10",
+	"20181-4-28T25:00:10",
+	"2018-10-218T04:00:10",
+	"2018-14-28T25:00:10",
+	"2018-142-8T25:00:10",
+	"2018-05-33T09:65:10.5",
+	"201",
+	"2018-011",
+	"2018-01-011",
+}
+
 func TestTypeForName(t *testing.T) {
 	for name, tid := range typeNameMap {
 		typ, ok := TypeForName(name)
@@ -46,22 +105,7 @@ func TestValueForType(t *testing.T) {
 }
 
 func TestParseTimeWithoutTZ(t *testing.T) {
-	tests := []struct {
-		in  string
-		out time.Time
-	}{
-		{in: "2018-10-28T04:00:10",
-			out: time.Date(2018, 10, 28, 4, 00, 10, 0, time.UTC)},
-		{in: "2018-05-30T09:30:10.5",
-			out: time.Date(2018, 5, 30, 9, 30, 10, 500000000, time.UTC)},
-		{in: "2018",
-			out: time.Date(2018, 1, 1, 0, 0, 0, 0, time.UTC)},
-		{in: "2018-01",
-			out: time.Date(2018, 1, 1, 0, 0, 0, 0, time.UTC)},
-		{in: "2018-01-01",
-			out: time.Date(2018, 1, 1, 0, 0, 0, 0, time.UTC)},
-	}
-	for _, tc := range tests {
+	for _, tc := range datesWithoutTz {
 		out, err := ParseTime(tc.in)
 		require.NoError(t, err)
 		require.EqualValues(t, tc.out, out)
@@ -75,26 +119,41 @@ func TestParseTimeWithTZ(t *testing.T) {
 	time.Local, err = time.LoadLocation("UTC")
 	require.NoError(t, err)
 
-	tests := []struct {
-		in  string
-		out time.Time
-	}{
-		{in: "2018-10-28T04:00:10Z",
-			out: time.Date(2018, 10, 28, 4, 00, 10, 0, time.UTC)},
-		{in: "2018-10-28T04:00:10-00:00",
-			out: time.Date(2018, 10, 28, 4, 00, 10, 0, time.UTC)},
-		{in: "2018-05-30T09:30:10.5Z",
-			out: time.Date(2018, 5, 30, 9, 30, 10, 500000000, time.UTC)},
-		{in: "2018-05-30T09:30:10.5-00:00",
-			out: time.Date(2018, 5, 30, 9, 30, 10, 500000000, time.UTC)},
-		{in: "2018-05-30T09:30:10-06:00",
-			out: time.Date(2018, 5, 30, 9, 30, 10, 0, time.FixedZone("", -6*60*60))},
-		{in: "2018-05-28T14:41:57+30:00",
-			out: time.Date(2018, 5, 28, 14, 41, 57, 0, time.FixedZone("", 30*60*60))},
-	}
-	for _, tc := range tests {
+	for _, tc := range datesWithTz {
 		out, err := ParseTime(tc.in)
 		require.NoError(t, err)
 		require.EqualValues(t, tc.out, out)
+	}
+}
+
+func TestParseTimeRejection(t *testing.T) {
+	var err error
+
+	// Set local time to UTC.
+	time.Local, err = time.LoadLocation("UTC")
+	require.NoError(t, err)
+
+	for _, invalidDate := range invalidDates {
+		_, err := ParseTime(invalidDate)
+		require.Error(t, err)
+	}
+}
+
+func BenchmarkParseTime(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		for _, tc := range datesWithTz {
+			ParseTime(tc.in)
+		}
+		for _, tc := range datesWithoutTz {
+			ParseTime(tc.in)
+		}
+	}
+}
+
+func BenchmarkParseTimeRejections(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		for _, invalidDate := range invalidDates {
+			ParseTime(invalidDate)
+		}
 	}
 }