benchmark                old ns/op     new ns/op     delta
BenchmarkParseTime-8     7103          4379          -38.35%

benchmark                old allocs     new allocs     delta
BenchmarkParseTime-8     39             16             -58.97%

benchmark                old bytes     new bytes     delta
BenchmarkParseTime-8     2144          896           -58.21%

<!-- Reviewable:start -->
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/dgraph/4693)
<!-- Reviewable:end -->
 
<!-- Dgraph:start -->
Docs Preview: [<img src="https://bl.ocks.org/prashant-shahi/raw/3a9f99bec84231cfe3c0e82cf883f159/0e588d908ad8c8b10958b87ebdd2ba68779ccf4f/dgraph.svg" height="34" align="absmiddle" alt="Dgraph Preview"/>](https://dgraph-cf298c6467-43108.surge.sh)
<!-- Dgraph:end -->