@mangalaman93 2020-01-29
@aakside Thank you for raising the PR. Will you check the CI build, seems like some of the tests fail. Also, will you provide the benchmark results with comparison using `benchcmp` tool?