@aakside 2020-01-29
@mangalaman93 
> @aakside Thank you for raising the PR. Will you check the CI build, seems like some of the tests fail. Also, will you provide the benchmark results with comparison using `benchcmp` tool?

Done, and edited first comment to include benchmark comparison