@aakside 2020-01-30
___
*[types/scalar_types.go, line 193 at r1](https://reviewable.io/reviews/dgraph-io/dgraph/4693#-LzsQfRo0xsacu2wZGDA:-LzsZB0V9ZDF2wtwDsQ9:b-rxm92t) ([raw file](https://github.com/dgraph-io/dgraph/blob/61e4c59f9cec9877c43a8c38dcc1ebce82ffb7cb/types/scalar_types.go#L193)):*
>If && is correct, then I don't think the first condition is doing anything.

Counterexample: `val = "2006-04"`
* Without the first condition, we would incorrectly proceed to Parse val as dateFormatY
* With the first condition, we would not do that, and would instead Parse val as dateFormatYM


<!-- Sent from Reviewable.io -->