@aakside 2020-02-05
___
*[types/scalar_types.go, line 197 at r2](https://reviewable.io/reviews/dgraph-io/dgraph/4693#-LzscUf696fTnP7DkiYq:-M-M_XZn5A5NXPg6dXh0:b98x36m) ([raw file](https://github.com/dgraph-io/dgraph/blob/c988329bd728bb5e0c036f4818fb766e0a8338de/types/scalar_types.go#L197)):*
<details><summary><i>Previously, manishrjain (Manish R Jain) wrote…</i></summary><blockquote>

Might check a dash here? Should optimize the rejection cases. In fact, the benchmark should be to check how fast can we reject strings which are not datetimes, but sort of look like them.

Could also look into building a larger regexp.
</blockquote></details>

Added a benchmark for rejections. I also tried checking for dashes and that made parse time slower so I left them out.


<!-- Sent from Reviewable.io -->