@aakside 2020-02-06
___
*[types/scalar_types.go, line 207 at r7](https://reviewable.io/reviews/dgraph-io/dgraph/4693#-M-QFLSs-Akxg0K7Ty4K:-M-R5xUt8p1N_rEN7Apc:bpb3vwk) ([raw file](https://github.com/dgraph-io/dgraph/blob/a6b27f70c6c16db3ed83a53b655305ccd5459ddd/types/scalar_types.go#L207)):*
<details><summary><i>Previously, manishrjain (Manish R Jain) wrote…</i></summary><blockquote>

I'll double check these assumptions with the Go folks in Slack.

Or, check for a dash at 4th place, and if it's there, then go one by one for RFC, YMDZone and dateTimeFormat.
</blockquote></details>

I verified this slightly revised check is correct by referring to https://tools.ietf.org/html/rfc3339#section-5.6 and testing time.Parse() against several negative cases (e.g., without an offset, without offset colon and 'Z', only hours in UTC offsets, etc.) to ensure it fails as expected.

(I haven't yet gotten a response in the Gophers slack #general channel)


<!-- Sent from Reviewable.io -->